Camera + handling blob
I fix the camera issue now it works. Also blob from our database now
works it saves a byte array and we can retrieve it by using the method
Curson.getBlob(columnIndex). And then we have to convert the byte array
to Bitmap i have done this in HistoryListViewAdapter in comments.