Filters
Add filters module on HistoryActivity. I added a menu item in actionBar
and when you press it FiltersDialog class begins. FiltersDialog extends
DialogFragment where there is a static listView with all the options for
filtering. Depend on the option there is another DialogFragment class
that allows you to select the attributes of the filter. To send data
from child DialogFragment to Parent DialogFragment i used an interface
in each child and the parent Fragment implements all the interfaces.
After a filter is chose we call the method refreshList from History
Activity to refresh the cursor.